requirements.txt
streamlit
python-dotenv
google-cloud-speech
google-cloud-texttospeech
moviepy
pydub
librosa
soundfile
requests